refactor: restore ExpenseLineTypeSwitch in is file
